Move recipe list and bottom signup components to homepage (#137)
* Move recipe list and bottom signup components to homepage

* Fix imports
